.history/backend/requirements_20251005182046.txt
Flask==3.0.0
flask-cors==4.0.0
PyMuPDF==1.23.8
opencv-python-headless==4.8.1.78
numpy==1.26.4
Pillow==10.1.0
google-cloud-vision==3.4.5
gunicorn==21.2.0